pyproject.toml
[project]
name = "rkllama"
version = "0.0.54"
authors = [
    { name="NotPunchnox", email="punchnoxpro@gmail.com" },
    { name="TomJacobsUK", email="tom@tomjacobs.co.uk" },
    { name="Daniel Ferreira", email="danielferreirandrade@gmail.com" },
]
description = "Ollama alternative for Rockchip NPU"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # Web framework (migrating Flask -> FastAPI)
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",

    # HTTP client
    "requests>=2.31.0",
    "aiohttp>=3.10.0",
    "aiofiles>=24.1.0",

    # ML/AI
    "huggingface_hub>=0.26.0",
    "transformers>=4.46.0",
    "torch>=2.0.0",
    "diffusers>=0.31.0",

    # Data validation
    "pydantic>=2.9.0",

    # Utilities
    "python-dotenv>=1.0.0",
    "psutil>=6.0.0",
    "numpy<2",

    # Image processing
    "pillow>=10.0.0",
    "opencv-python>=4.10.0",

    # Audio processing
    "piper-tts==1.3.0",
    "pydub>=0.25.0",
    "soxr>=0.5.0",
    "soundfile>=0.12.0",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "Framework :: FastAPI",
]
license = "GPL-3.0"
license-files = ["LICEN[CS]E*"]

[project.optional-dependencies]
# RKNN toolkit for ARM64 platforms
rknn = [
    "rknn-toolkit-lite2 ; python_version == '3.12'",
    "rknn-toolkit-lite2 ; python_version == '3.11'",
    "rknn-toolkit-lite2 ; python_version == '3.10'",
]
# S3 support for model pulling
s3 = [
    "aioboto3>=13.0.0",
    "boto3>=1.35.0",
]
# Development dependencies
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.27.0",
    "respx>=0.21.0",
    "ruff>=0.8.0",
    "mypy>=1.14.0",
    "pre-commit>=4.0.0",
]
# All optional dependencies
all = [
    "rkllama[rknn,s3,dev]",
]

[tool.uv.sources]
rknn-toolkit-lite2 = [
    { path = "src/rkllama/lib/rknn_toolkit_lite2-2.3.2-cp312-cp312-manylinux_2_17_aarch64.manylinux2014_aarch64.whl", marker = "python_version == '3.12'" },
    { path = "src/rkllama/lib/rknn_toolkit_lite2-2.3.2-cp311-cp311-manylinux_2_17_aarch64.manylinux2014_aarch64.whl", marker = "python_version == '3.11'" },
    { path = "src/rkllama/lib/rknn_toolkit_lite2-2.3.2-cp310-cp310-manylinux_2_17_aarch64.manylinux2014_aarch64.whl", marker = "python_version == '3.10'" },
]


[tool.setuptools]
include-package-data = true
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"rkllama.lib" = ["*.so", "*.sh"]

#[tool.setuptools.data-files]
#"rkllama/config/lib" = [
#    "lib/*"
#]

[project.urls]
Homepage = "https://github.com/NotPunchnox/rkllama"
Issues = "https://github.com/NotPunchnox/rkllama/issues"

[build-system]
requires = ["setuptools>=77.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project.scripts]
rkllama_server = "rkllama.server.app:main"
rkllama_client = "rkllama.client.client:main"

# ============================================================================
# Tool Configuration
# ============================================================================

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.27.0",
    "respx>=0.21.0",
    "ruff>=0.8.0",
    "mypy>=1.14.0",
    "pre-commit>=4.0.0",
]

[tool.ruff]
target-version = "py310"
line-length = 120
src = ["src"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "B904",   # raise without from inside except
    "ARG001", # unused function argument
]

[tool.ruff.lint.isort]
known-first-party = ["rkllama"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.coverage.run]
source = ["src/rkllama"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
